# Reduce risk of server closing conn (slow consumer) if sending to itself