When a client sends data and has at least a matching subscriber,
there may be a case where the flusher is holding the connection's
lock while trying to flush the write buffer, which cannot complete
because the server is not draining data at the moment since it is
itself trying to send data to the client, which processMsg() function
is waiting on the connection's lock. After 2 seconds, the server
will close the client connection.
Using a dedicated lock for the subscriptions map that processMsg
needs to access is reducing the risk of lock contention between
publish/flusher and the readloop's processMsg function.
There would still be a case where if the subscription in the client
is considered a slow consumer and messages are dropped, we need to
acquire the connection lock there. But in cases where subscribers
keep up, we would reduce the risk of server's slow consumer error.