@kozlovic 2017-04-24
@derekcollison The main goal is to avoid the server cutting off the connection (slow consumer), but the side effect is an improvement in performance for the PubSub using same connection:
```
Ivans-MacBook-Pro:go-nats ivan$ benchcmp old.txt new.txt 
benchmark                                     old ns/op     new ns/op     delta
BenchmarkPublishSpeed-8                       125           129           +3.20%
BenchmarkPubSubSpeed-8                        1495          579           -61.27%
BenchmarkAsyncSubscriptionCreationSpeed-8     28716         29348         +2.20%
BenchmarkSyncSubscriptionCreationSpeed-8      43834         33926         -22.60%
BenchmarkInboxCreation-8                      122           123           +0.82%
BenchmarkRequest-8                            67531         67824         +0.43%
BenchmarkPublishSpeedViaChan-8                1667          1697          +1.80%

benchmark                                     old allocs     new allocs     delta
BenchmarkAsyncSubscriptionCreationSpeed-8     14             14             +0.00%
BenchmarkSyncSubscriptionCreationSpeed-8      13             13             +0.00%
BenchmarkRequest-8                            58             58             +0.00%

benchmark                                     old bytes     new bytes     delta
BenchmarkAsyncSubscriptionCreationSpeed-8     1326          1325          -0.08%
BenchmarkSyncSubscriptionCreationSpeed-8      66301         66299         -0.00%
BenchmarkRequest-8                            2666          2666          +0.00%
```